Use `getdescendant` for `get_node_at_location`
Getting the node at a certain location in a tree is
currently allocating memory on the heap.

As this is a common operation, it is likely beneficial to
get rid of all allocations, if possible. `getdescendant`
from `AbstractTrees` performs the same operation as
`get_node_at_location`, so this change is fairly
straightforward. For the `Hole` method of `get_node_at_location`,
the comparison to `[]` was also allocating and switching
this to `isempty(location)` avoids the allocation.

I've relaxed both signatures so that location can take
something other than a `Vector{Int}`. This allows for
use of a `SubArray` or other `<:AbstractVector` types
as a `location`.